Safari-609.1.11.1/LayoutTests/transforms/2d/computed-style-origin-expected.txt

This tests computed style values for transform-origin and perspective-origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test1Style.transformOrigin is '10px 10px'
PASS test2Style.transformOrigin is '10px 10px'
PASS test3Style.transformOrigin is '0px 20px'
PASS test4Style.transformOrigin is '5px 7px'
PASS test5Style.transformOrigin is '15px 15px'
PASS test1Style.perspectiveOrigin is '10px 10px'
PASS test2Style.perspectiveOrigin is '10px 10px'
PASS test3Style.perspectiveOrigin is '0px 20px'
PASS test4Style.perspectiveOrigin is '5px 7px'
PASS test5Style.perspectiveOrigin is '15px 15px'
PASS test6Style.perspectiveOrigin is '50px 50px'

PASS successfullyParsed is true

TEST COMPLETE
